Attributes tab was not functioning correctly in case only one language was installed

+++ public/javascript/backend/Template.js
@@ -3,48 +3,48 @@
  *	@author Integry Systems
  */
 Backend.Template = Class.create();
-Backend.Template.prototype = 
+Backend.Template.prototype =
 {
   	treeBrowser: null,
-  	
+
   	urls: new Array(),
-	  
+
 	initialize: function(categories)
 	{
 		this.treeBrowser = new dhtmlXTreeObject("templateBrowser","","", false);
         Backend.Breadcrumb.setTree(this.treeBrowser);
-		
+
 		this.treeBrowser.def_img_x = 'auto';
 		this.treeBrowser.def_img_y = 'auto';
-				
+
 		this.treeBrowser.setImagePath("image/backend/dhtmlxtree/");
 		this.treeBrowser.setOnClickHandler(this.activateCategory.bind(this));
 
-		this.treeBrowser.showFeedback = 
-			function(itemId) 
+		this.treeBrowser.showFeedback =
+			function(itemId)
 			{
 				if (!this.iconUrls)
 				{
-					this.iconUrls = new Object();	
+					this.iconUrls = new Object();
 				}
-				
+
 				this.iconUrls[itemId] = this.getItemImage(itemId, 0, 0);
 				this.setItemImage(itemId, '../../../image/indicator.gif');
 			}
-		
-		this.treeBrowser.hideFeedback = 
+
+		this.treeBrowser.hideFeedback =
 			function()
 			{
 				for (var itemId in this.iconUrls)
 				{
-					this.setItemImage(itemId, this.iconUrls[itemId]);	
-				}				
+					this.setItemImage(itemId, this.iconUrls[itemId]);
+				}
 			}
-		
-    	this.insertTreeBranch(categories, 0);    
+
+    	this.insertTreeBranch(categories, 0);
     	this.treeBrowser.closeAllItems();
 	},
-	
+
 	insertTreeBranch: function(treeBranch, rootId)
 	{
 		for (k in treeBranch)
@@ -52,15 +52,15 @@ Backend.Template.prototype =
 		  	if('function' != typeof treeBranch[k])
 		  	{
 				this.treeBrowser.insertNewItem(rootId, treeBranch[k].id, k, null, 0, 0, 0, '');
-				
+
 				if (treeBranch[k].subs)
 				{
 					this.insertTreeBranch(treeBranch[k].subs, treeBranch[k].id);
 				}
 			}
-		}  	
-	},    
-	
+		}
+	},
+
 	activateCategory: function(id)
 	{
         if (!this.treeBrowser.hasChildren(id))
@@ -77,24 +77,27 @@ Backend.Template.prototype =
 			if ($('body'))
 			{
 				editAreaLoader.delete_instance("body");
-				
-        		var langs = $('templateContent').down('.languageFormContent').getElementsByTagName('textarea');
-        		for (k = 0; k < langs.length; k++)
-        		{
-                    if ($('frame_' + langs[k].id))
-                    {
-                        editAreaLoader.delete_instance(langs[k].id);
-                    }
-        		}				
+
+				if ($('templateContent').down('.languageFormContent'))
+				{
+					var langs = $('templateContent').down('.languageFormContent').getElementsByTagName('textarea');
+					for (k = 0; k < langs.length; k++)
+					{
+						if ($('frame_' + langs[k].id))
+						{
+							editAreaLoader.delete_instance(langs[k].id);
+						}
+					}
+				}
 			}
 		}
-	},	
-	
+	},
+
 	displayTemplate: function(response)
 	{
 		this.treeBrowser.hideFeedback();
 		Event.observe($('cancel'), 'click', this.cancel.bindAsEventListener(this));
-		
+
 		if ($('code'))
 		{
 			new Backend.TemplateHandler($('templateForm'));
@@ -102,28 +105,28 @@ Backend.Template.prototype =
 		else
 		{
 			new Backend.EmailTemplateHandler($('templateForm'));
-		}		
+		}
 	},
 
     cancel: function()
     {
-		new LiveCart.AjaxUpdater(this.urls['empty'], 'templateContent', 'settingsIndicator');        
-    }	
+		new LiveCart.AjaxUpdater(this.urls['empty'], 'templateContent', 'settingsIndicator');
+    }
 }
 
 /**
  *  Template editor form handler
  */
 Backend.TemplateHandler = Class.create();
-Backend.TemplateHandler.prototype = 
+Backend.TemplateHandler.prototype =
 {
 	form: null,
-	
+
 	initialize: function(form)
 	{
 		this.form = form;
 		this.form.onsubmit = this.submit.bindAsEventListener(this);
-		
+
 		editAreaLoader.init({
 			id : "code",		// textarea id
 			syntax: "html",			// syntax to be uses for highgliting
@@ -132,23 +135,23 @@ Backend.TemplateHandler.prototype =
 			allow_resize: true
 			}
 		);
-		
+
 		// set cursor at the first line
-		editAreaLoader.setSelectionRange('code', 0, 0);		
+		editAreaLoader.setSelectionRange('code', 0, 0);
 	},
-	
+
 	submit: function()
 	{
         $('code').value = editAreaLoader.getValue('code');
 		new LiveCart.AjaxRequest(this.form, null, this.saveComplete.bind(this));
 		return false;
 	},
-	
+
 	saveComplete: function(originalRequest)
 	{
 		if (opener)
 		{
-            opener.location.reload();	            
+            opener.location.reload();
         }
 	}
 }
@@ -157,10 +160,10 @@ Backend.TemplateHandler.prototype =
  *  E-mail template editor form handler
  */
 Backend.EmailTemplateHandler = Class.create();
-Backend.EmailTemplateHandler.prototype = 
+Backend.EmailTemplateHandler.prototype =
 {
 	form: null,
-	
+
 	initialize: function(form)
 	{
 		this.form = form;
@@ -176,46 +179,52 @@ Backend.EmailTemplateHandler.prototype =
 		);
 
 		// set cursor at the first line
-		editAreaLoader.setSelectionRange('body', 0, 0);		
+		editAreaLoader.setSelectionRange('body', 0, 0);
 
 		// initialize editors for other languages
-		var langs = $('templateContent').down('.languageFormTabs').getElementsByTagName('li');
-		for (k = 0; k < langs.length; k++)
+		if ($('templateContent').down('.languageFormContent'))
 		{
-			Event.observe(langs[k], 'click',
-				function(e)
-				{
-					var lang = this.className.match(/Tabs_([a-z]{2})/)[1];
-					var textarea = $('templateContent').down('.languageFormContent').down('.languageFormContainer_' + lang).down('textarea');
+			var langs = $('templateContent').down('.languageFormTabs').getElementsByTagName('li');
+			for (k = 0; k < langs.length; k++)
+			{
+				Event.observe(langs[k], 'click',
+					function(e)
+					{
+						var lang = this.className.match(/Tabs_([a-z]{2})/)[1];
+						var textarea = $('templateContent').down('.languageFormContent').down('.languageFormContainer_' + lang).down('textarea');
 
-					editAreaLoader.init({
-						id : textarea.id,		// textarea id
-						syntax: "html",			// syntax to be uses for highgliting
-						start_highlight: true,		// to display with highlight mode on start-up
-						allow_toggle: false,
-						allow_resize: true
-						}
-					);		
+						editAreaLoader.init({
+							id : textarea.id,		// textarea id
+							syntax: "html",			// syntax to be uses for highgliting
+							start_highlight: true,		// to display with highlight mode on start-up
+							allow_toggle: false,
+							allow_resize: true
+							}
+						);
 
-				}
-			);
+					}
+				);
+			}
 		}
 	},
-	
+
 	submit: function()
 	{
         $('body').value = editAreaLoader.getValue('body');
 
-		var langs = $('templateContent').down('.languageFormContent').getElementsByTagName('textarea');
-		for (k = 0; k < langs.length; k++)
+		if ($('templateContent').down('.languageFormContent'))
 		{
-			langs[k].value = editAreaLoader.getValue(langs[k].id);
+			var langs = $('templateContent').down('.languageFormContent').getElementsByTagName('textarea');
+			for (k = 0; k < langs.length; k++)
+			{
+				langs[k].value = editAreaLoader.getValue(langs[k].id);
+			}
 		}
-	
+
 		new LiveCart.AjaxRequest(this.form, null, this.saveComplete.bind(this));
 		return false;
 	},
-	
+
 	saveComplete: function(originalRequest)
 	{
 	}
@@ -226,8 +235,8 @@ function decode64(inp)
 
 var keyStr = "ABCDEFGHIJKLMNOPQRSTUVWXYZ" + //all caps
 "abcdefghijklmnopqrstuvwxyz" + //all lowercase
-"0123456789+/="; 
- 
+"0123456789+/=";
+
 var out = ""; //This is the output
 var chr1, chr2, chr3 = ""; //These are the 3 decoded bytes
 var enc1, enc2, enc3, enc4 = ""; //These are the 4 bytes to be decoded
@@ -243,7 +252,7 @@ alert("There were invalid base64 characters in the input text.\n" +
 }
 inp = inp.replace(/[^A-Za-z0-9\+\/\=]/g, "");
 
-do { //Here’s the decode loop.
+do { //Hereï¿½s the decode loop.
 
 //Grab 4 bytes of encoded content.
 enc1 = keyStr.indexOf(inp.charAt(i++));
@@ -251,7 +260,7 @@ enc2 = keyStr.indexOf(inp.charAt(i++));
 enc3 = keyStr.indexOf(inp.charAt(i++));
 enc4 = keyStr.indexOf(inp.charAt(i++));
 
-//Heres the decode part. There’s really only one way to do it.
+//Heres the decode part. Thereï¿½s really only one way to do it.
 chr1 = (enc1 << 2) | (enc2 >> 4);
 chr2 = ((enc2 & 15) << 4) | (enc3 >> 2);
 chr3 = ((enc3 & 3) << 6) | enc4;
@@ -277,33 +286,33 @@ enc1 = enc2 = enc3 = enc4 = "";
 return _utf8_decode(out);
 }
 
- // private method for UTF-8 decoding  
-function _utf8_decode(utftext) {  
-     var string = "";  
-     var i = 0;  
-     var c, c1, c2 = 0;  
+ // private method for UTF-8 decoding
+function _utf8_decode(utftext) {
+     var string = "";
+     var i = 0;
+     var c, c1, c2 = 0;
 
-     while ( i < utftext.length ) {  
+     while ( i < utftext.length ) {
 
-         c = utftext.charCodeAt(i);  
+         c = utftext.charCodeAt(i);
 
-         if (c < 128) {  
-             string += String.fromCharCode(c);  
-             i++;  
-         }  
-         else if((c > 191) && (c < 224)) {  
-             c2 = utftext.charCodeAt(i+1);  
-             string += String.fromCharCode(((c & 31) << 6) | (c2 & 63));  
-             i += 2;  
-         }  
-         else {  
-             c2 = utftext.charCodeAt(i+1);  
-             c3 = utftext.charCodeAt(i+2);  
-             string += String.fromCharCode(((c & 15) << 12) | ((c2 & 63) << 6) | (c3 & 63));  
-             i += 3;  
-         }  
+         if (c < 128) {
+             string += String.fromCharCode(c);
+             i++;
+         }
+         else if((c > 191) && (c < 224)) {
+             c2 = utftext.charCodeAt(i+1);
+             string += String.fromCharCode(((c & 31) << 6) | (c2 & 63));
+             i += 2;
+         }
+         else {
+             c2 = utftext.charCodeAt(i+1);
+             c3 = utftext.charCodeAt(i+2);
+             string += String.fromCharCode(((c & 15) << 12) | ((c2 & 63) << 6) | (c3 & 63));
+             i += 3;
+         }
 
-     }  
+     }
 
-     return string;  
-}  
+     return string;
+}
